+++ java-api/src/main/java/com/sap/cloud/security/token/ProviderNotFoundException.java
@@ -1,0 +1,16 @@
+package com.sap.cloud.security.token;
+
+/**
+ * A dedicated runtime exception for missing implementations in
+ * {@link java.util.ServiceLoader} context
+ */
+public class ProviderNotFoundException extends RuntimeException {
+
+	public ProviderNotFoundException() {
+		super();
+	}
+
+	public ProviderNotFoundException(String message) {
+		super(message);
+	}
+}