+++ java-api/src/test/java/com/sap/cloud/security/token/test/CustomTokenFactory.java
@@ -1,0 +1,12 @@
+package com.sap.cloud.security.token.test;
+
+import com.sap.cloud.security.token.Token;
+import com.sap.cloud.security.token.TokenFactory;
+import org.mockito.Mockito;
+
+public class CustomTokenFactory implements TokenFactory {
+	@Override
+	public Token create(String jwtToken) {
+		return Mockito.mock(Token.class);
+	}
+}