+++ spring-xsuaa-test/src/test/java/com/sap/cloud/security/xsuaa/test/JwtGeneratorTest.java
@@ -5,17 +5,15 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.sap.cloud.security.xsuaa.test.JwtGenerator.TokenClaims;
 import com.sap.cloud.security.xsuaa.test.JwtGenerator.TokenHeaders;
+import org.json.JSONArray;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.security.oauth2.jwt.Jwt;
-
-import net.minidev.json.JSONArray;
 
 public class JwtGeneratorTest {
 	private JwtGenerator jwtGenerator;
@@ -34,7 +32,8 @@ public class JwtGeneratorTest {
 		assertThat(jwt.getClaimAsString("zid"), equalTo(JwtGenerator.DEFAULT_IDENTITY_ZONE_ID));
 		assertThat(jwt.getExpiresAt(), not(equals(nullValue())));
 		assertThat(jwt.getAudience(), is(nullValue()));
-		assertThat(getExternalAttributeFromClaim(jwt, "zdn"), isEmptyString());
+		assertThat(getExternalAttributeFromClaim(jwt, "zdn"), is(emptyString()));
+		assertThat(getExternalAttributeFromClaim(jwt, "enhancer"), equalTo("XSUAA"));
 	}
 
 	@Test
@@ -67,7 +66,7 @@ public class JwtGeneratorTest {
 
 	@Test
 	public void testTokenWithScopes() {
-		Jwt jwt = jwtGenerator.addScopes(new String[] { DUMMY_SCOPE, ANOTHER_SCOPE }).getToken();
+		Jwt jwt = jwtGenerator.addScopes(DUMMY_SCOPE, ANOTHER_SCOPE).getToken();
 		assertThat(jwt.getClaimAsStringList("scope"), hasItems(DUMMY_SCOPE, ANOTHER_SCOPE));
 	}
 
@@ -77,8 +76,8 @@ public class JwtGeneratorTest {
 				.addAttribute(ANOTHER_ATTRIBUTE, new String[] { ANOTHER_ATTRIBUTE_VALUE, ANOTHER_ATTRIBUTE_VALUE_2 })
 				.getToken();
 		Map<String, Object> attributes = jwt.getClaimAsMap("xs.user.attributes");
-		assertThat((JSONArray) attributes.get(DUMMY_ATTRIBUTE), contains(DUMMY_ATTRIBUTE));
-		assertThat((JSONArray) attributes.get(ANOTHER_ATTRIBUTE),
+		assertThat(new JSONArray((ArrayList) attributes.get(DUMMY_ATTRIBUTE)), contains(DUMMY_ATTRIBUTE));
+		assertThat(new JSONArray((ArrayList) attributes.get(ANOTHER_ATTRIBUTE)),
 				contains(ANOTHER_ATTRIBUTE_VALUE, ANOTHER_ATTRIBUTE_VALUE_2));
 	}
 
@@ -103,8 +102,7 @@ public class JwtGeneratorTest {
 		assertThat(jwt.getClaimAsStringList("scope"), hasItems("openid", "testScope", "testApp.localScope"));
 
 		Map<String, Object> attributes = jwt.getClaimAsMap("xs.user.attributes");
-		assertThat((JSONArray) attributes.get("usrAttr"), contains("value_1", "value_2"));
-		jwt.getTokenValue();
+		assertThat(new JSONArray((ArrayList) attributes.get("usrAttr")), contains("value_1", "value_2"));
 	}
 
 	@Test